Implemented JWT-based authentication for enhanced security and user authorization

+++ controllers/todoController.js
@@ -2,16 +2,16 @@ const asyncHandler = require('express-async-handler');
 const Todo = require('../models/todoModel');
 
 //@desc Get all todos
-//@route GET /api/contact
-//@access public
+//@route GET /api/todos
+//@access private
 const getTodos = asyncHandler(async (req, res) => {
-    const todo = await Todo.find({});
+    const todo = await Todo.find({user_id: req.user.id});
     res.status(200).json(todo);
 });
 
 //@desc Create New todos
-//@route POST /api/contact
-//@access public
+//@route POST /api/todos
+//@access private
 const createTodo = asyncHandler(async (req, res) => {
     const {todo, priority, status, category, dueDate} = req.body;
     if (!todo || !priority || !status || !category || !dueDate) {
@@ -23,14 +23,15 @@ const createTodo = asyncHandler(async (req, res) => {
         priority,
         status,
         category,
-        dueDate
+        dueDate,
+        user_id: req.user.id
     });
     res.status(201).json(newTodo);
 });
 
 //@desc Get todo
-//@route GET /api/contact/:id
-//@access public
+//@route GET /api/todos/:id
+//@access private
 const getTodo = asyncHandler(async (req, res) => {
     const { id } = req.params;
     const todo = await Todo.findById(id);
@@ -38,12 +39,18 @@ const getTodo = asyncHandler(async (req, res) => {
         res.status(404);
         throw new Error("Todo not found");
     }
+
+    if (todo.user_id.toString() !== req.user.id) {
+        res.status(403);
+        throw new Error("User don't have permission to update other user todo");
+    }
+
     res.status(200).json(todo);
 });
 
 //@desc Update todo
-//@route PUT /api/contact/:id
-//@access public
+//@route PUT /api/todos/:id
+//@access private
 const updateTodo = asyncHandler(async (req, res) => {
     const { id } = req.params;
     
@@ -52,14 +59,19 @@ const updateTodo = asyncHandler(async (req, res) => {
         res.status(404);
         throw new Error("Todo not found");
     }
+
+    if (todo.user_id.toString() !== req.user.id) {
+        res.status(403);
+        throw new Error("User don't have permission to update other user todo");
+    }
     
     const updatedTodo = await Todo.findByIdAndUpdate(id, req.body);
     res.status(200).json(updatedTodo);
 });
 
 //@desc Delete todo
-//@route DELETE /api/contact/:id
-//@access public
+//@route DELETE /api/todos/:id
+//@access private
 const deleteTodo = asyncHandler(async (req, res) => {
     const { id } = req.params;
     
@@ -67,6 +79,11 @@ const deleteTodo = asyncHandler(async (req, res) => {
     if (!todo) {
         res.status(404);
         throw new Error("Todo not found");
+    }
+
+    if (todo.user_id.toString() !== req.user.id) {
+        res.status(403);
+        throw new Error("User don't have permission to delete other user todo");
     }
     
     const deletedTodo = await Todo.findByIdAndDelete(id);
